requirements.txt
# Core dependencies
click>=8.0.0
colorama>=0.4.4
prompt-toolkit>=3.0.0
rich>=13.0.0

# AI/ML dependencies
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
huggingface-hub>=0.17.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
psutil>=5.9.0

# Natural Language Processing
spacy>=3.6.0
nltk>=3.8.0

# HTTP requests
requests>=2.31.0
httpx>=0.24.0

# Configuration
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Async support
asyncio
aiohttp>=3.8.0

# Logging
loguru>=0.7.0

# File operations
pathlib
watchdog>=3.0.0

# Security
cryptography>=41.0.0

# Testing (dev dependencies)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
coverage>=7.2.0

# Code quality (dev dependencies)
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0

# Documentation (dev dependencies)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Build tools
setuptools>=68.0.0
wheel>=0.40.0 